scenarios: refactor types
